Have our own Variable
Subclass Pymbolic.variable so that we can control the color of output. Consequently refactor mappers into another module to avoid circular dependencies.
